ajustado o código de acordo com o do vídeo

+++ pooAula10/src/Professor.java
@@ -1,8 +1,8 @@
 public class Professor extends Pessoa{
     private String especialidade;
-    private int salario;
+    private float salario;
 
-    public Professor(String nome, int idade, String sexo, String especialidade, int salario){
+    public Professor(String nome, int idade, String sexo, String especialidade, float salario){
         super(nome, idade, sexo);
         this.especialidade = especialidade;
         this.salario = salario;
@@ -20,11 +20,11 @@ public class Professor extends Pessoa{
         this.especialidade = especialidade;
     }
 
-    public int getSalario() {
+    public float getSalario() {
         return salario;
     }
 
-    public void setSalario(int salario) {
+    public void setSalario(float salario) {
         this.salario = salario;
     }
 }
